🌱 remove dead maturity detector
