+++ changelog.d/changed-5780-remove-maturity-detector.md
@@ -1,0 +1,1 @@
+- Removed the unused knowledge maturity detector and its self-contained tests so the package no longer exposes a dead repository-maturity analysis surface.
