+++ src/pkg/knowledge/test_helpers_test.go
@@ -1,0 +1,19 @@
+package knowledge
+
+import (
+	"log/slog"
+	"net/http/httptest"
+	"os"
+
+	gh "github.com/google/go-github/v72/github"
+)
+
+func knowledgeTestLogger() *slog.Logger {
+	return slog.New(slog.NewTextHandler(os.Stderr, &slog.HandlerOptions{Level: slog.LevelError}))
+}
+
+func ghClientFromServer(server *httptest.Server) *gh.Client {
+	client := gh.NewClient(nil)
+	client.BaseURL, _ = client.BaseURL.Parse(server.URL + "/")
+	return client
+}
